+++ NEWS.md
@@ -1,8 +1,9 @@
 
-'replyr' 0.4.1 2017-07-05
+'replyr' 0.4.2 2017-07-05
 
  * fix spelling of tableDescription()
  * add diagrams to join controller.
+ * prepare for sparklyr rbind().
 
 'replyr' 0.4.0 2017-06-12